projects page complete
