+++ src/pages/Projects/Project.css
@@ -1,0 +1,41 @@
+.projectsPage{
+    display: flex;
+    flex-direction: row;
+    flex-wrap: wrap;
+    justify-content: center;
+    align-items: center;
+    gap: 10px;
+    margin-top: 1200px;
+    margin-bottom: 200px;
+}
+.card-projects{
+    display: flex;
+    flex-direction: row;
+    flex-wrap: wrap;
+    justify-content: center;
+    align-items: center;
+    text-align: left;
+    max-width: 900px;
+    gap: 20px;
+}
+.project-image{
+    /* max-width: 300px; */
+    max-height: 300px;
+    /* justify-content: center;
+    align-items: center;
+    text-align: left; */
+    left: 0;
+    flex: 1;
+    border-radius: 10px;
+}
+.projectContent{
+    flex: 3;
+    color: #4b423b;
+}
+.link > a{
+    text-decoration: none;
+    color: #4b423b;
+}
+.project-desc{
+    line-height: 30px;
+}
